+++ content/index.js
@@ -10,6 +10,7 @@ const Document = require("./document");
 const { getPopularities } = require("./popularities");
 const Redirect = require("./redirect");
 const Image = require("./image");
+const Archive = require("./archive");
 const { buildURL, memoize, slugToFolder, execGit } = require("./utils");
 const { resolveFundamental } = require("../libs/fundamental-redirects");
 
@@ -26,6 +27,7 @@ module.exports = {
   Document,
   Redirect,
   Image,
+  Archive,
 
   buildURL,
   memoize,
